+++ Packages/com.mimylab.fukuroudon/Runtime/ActiveRelay/Scripts/ActiveRelayToEffect.cs
@@ -11,23 +11,39 @@ namespace MimyLab.FukuroUdon
     //using VRC.SDKBase;
     //using VRC.Udon;
 
+    public enum ActiveRelayToEffectAudioSourceState
+    {
+        NoChange = default,
+        Play,
+        Pause,
+        Stop
+    }
+
+    public enum ActiveRelayToEffectEmissionModule
+    {
+        NoChange = default,
+        Enable,
+        Disable
+    }
+
     [Icon(ComponentIconPath.FukuroUdon)]
     [AddComponentMenu("Fukuro Udon/Active Relay/ActiveRelay to Effect")]
     [UdonBehaviourSyncMode(BehaviourSyncMode.None)]
     public class ActiveRelayToEffect : UdonSharpBehaviour
     {
-        [Tooltip("Only particle emission module toggle")]
-        [SerializeField]
-        private bool _invert = false;
-
         [Header("Settings when active")]
         [SerializeField]
         private AudioSource _audioSourceForActive = null;
         [SerializeField]
+        private ActiveRelayToEffectAudioSourceState _audioSourceStateForActive = default;
+        [SerializeField]
         private AudioClip _soundForActive = null;
 
+        [Space]
         [SerializeField]
         private ParticleSystem _particleForActive = null;
+        [SerializeField]
+        private ActiveRelayToEffectEmissionModule _emissionModuleForActive = default;
         [SerializeField]
         private int _emitForActive = 0;
 
@@ -35,10 +51,15 @@ namespace MimyLab.FukuroUdon
         [SerializeField]
         private AudioSource _audioSourceForInactive = null;
         [SerializeField]
+        private ActiveRelayToEffectAudioSourceState _audioSourceStateForInactive = default;
+        [SerializeField]
         private AudioClip _soundForInactive = null;
 
+        [Space]
         [SerializeField]
         private ParticleSystem _particleForInactive = null;
+        [SerializeField]
+        private ActiveRelayToEffectEmissionModule _emissionModuleForInactive = default;
         [SerializeField]
         private int _emitForInactive = 0;
 
@@ -46,19 +67,31 @@ namespace MimyLab.FukuroUdon
         {
             if (_audioSourceForActive)
             {
-                _audioSourceForActive.PlayOneShot(_soundForActive);
+                switch (_audioSourceStateForActive)
+                {
+                    case ActiveRelayToEffectAudioSourceState.Play: _audioSourceForActive.Play(); break;
+                    case ActiveRelayToEffectAudioSourceState.Pause: _audioSourceForActive.Pause(); break;
+                    case ActiveRelayToEffectAudioSourceState.Stop: _audioSourceForActive.Stop(); break;
+                }
+
+                if (_soundForActive)
+                {
+                    _audioSourceForActive.PlayOneShot(_soundForActive);
+                }
             }
 
             if (_particleForActive)
             {
+                var emission = _particleForActive.emission;
+                switch (_emissionModuleForActive)
+                {
+                    case ActiveRelayToEffectEmissionModule.Enable: emission.enabled = true; break;
+                    case ActiveRelayToEffectEmissionModule.Disable: emission.enabled = false; break;
+                }
+
                 if (_emitForActive > 0)
                 {
                     _particleForActive.Emit(_emitForActive);
-                }
-                else
-                {
-                    var emission = _particleForActive.emission;
-                    emission.enabled = !_invert;
                 }
             }
         }
@@ -67,19 +100,31 @@ namespace MimyLab.FukuroUdon
         {
             if (_audioSourceForInactive)
             {
-                _audioSourceForInactive.PlayOneShot(_soundForInactive);
+                switch (_audioSourceStateForInactive)
+                {
+                    case ActiveRelayToEffectAudioSourceState.Play: _audioSourceForInactive.Play(); break;
+                    case ActiveRelayToEffectAudioSourceState.Pause: _audioSourceForInactive.Pause(); break;
+                    case ActiveRelayToEffectAudioSourceState.Stop: _audioSourceForInactive.Stop(); break;
+                }
+
+                if (_soundForInactive)
+                {
+                    _audioSourceForInactive.PlayOneShot(_soundForInactive);
+                }
             }
 
             if (_particleForInactive)
             {
+                var emission = _particleForInactive.emission;
+                switch (_emissionModuleForInactive)
+                {
+                    case ActiveRelayToEffectEmissionModule.Enable: emission.enabled = true; break;
+                    case ActiveRelayToEffectEmissionModule.Disable: emission.enabled = false; break;
+                }
+
                 if (_emitForInactive > 0)
                 {
                     _particleForInactive.Emit(_emitForInactive);
-                }
-                else
-                {
-                    var emission = _particleForInactive.emission;
-                    emission.enabled = _invert;
                 }
             }
         }